+++ services/api/server/config.go
@@ -4,10 +4,8 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"net/http"
-	"strings"
 
 	"github.com/go-redis/redis/v7"
-	nickHelix "github.com/nicklaw5/helix"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -102,30 +100,4 @@ func writeJSON(w http.ResponseWriter, data interface{}, code int) {
 	if err != nil {
 		log.Errorf("Faile to writeJSON: %s", err)
 	}
-}
-
-func (s *Server) authenticate(r *http.Request) (bool, *nickHelix.ValidateTokenResponse) {
-	scToken := strings.TrimPrefix(r.Header.Get("Authorization"), "Bearer ")
-
-	val, err := s.store.Client.HGet("accessTokens", scToken).Result()
-	if err != nil {
-		log.Error(err)
-		return false, nil
-	}
-
-	var accessToken *nickHelix.UserAccessTokenResponse
-	if err := json.Unmarshal([]byte(val), &accessToken); err != nil {
-		log.Error(err)
-		return false, nil
-	}
-
-	success, resp, err := s.helixClient.Client.ValidateToken(accessToken.Data.AccessToken)
-	if !success || err != nil {
-		if err != nil {
-			log.Error(err)
-		}
-		return false, nil
-	}
-
-	return success, resp
 }
