+++ services/api/server/oauth.go
@@ -5,9 +5,18 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"strings"
+
+	nickHelix "github.com/nicklaw5/helix"
 
 	log "github.com/sirupsen/logrus"
 )
+
+type userAcessTokenData struct {
+	AccessToken  string
+	RefreshToken string
+	Scope        string
+}
 
 func (s *Server) handleCallback(w http.ResponseWriter, r *http.Request) {
 	code := r.URL.Query().Get("code")
@@ -19,7 +28,7 @@ func (s *Server) handleCallback(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	marshalled, err := json.Marshal(resp.Data)
+	marshalled, err := json.Marshal(userAcessTokenData{resp.Data.AccessToken, resp.Data.RefreshToken, strings.Join(resp.Data.Scopes, " ")})
 	if err != nil {
 		log.Error(err)
 		s.dashboardRedirect(w, r, http.StatusInternalServerError, "")
@@ -43,4 +52,72 @@ func (s *Server) dashboardRedirect(w http.ResponseWriter, r *http.Request, statu
 	}
 
 	http.Redirect(w, r, s.cfg.WebBaseUrl+"/dashboard"+"?"+params.Encode(), http.StatusFound)
+}
+
+func (s *Server) authenticate(r *http.Request) (bool, *nickHelix.ValidateTokenResponse) {
+	scToken := strings.TrimPrefix(r.Header.Get("Authorization"), "Bearer ")
+
+	val, err := s.store.Client.HGet("accessTokens", scToken).Result()
+	if err != nil {
+		log.Error(err)
+		return false, nil
+	}
+
+	var token userAcessTokenData
+	if err := json.Unmarshal([]byte(val), &token); err != nil {
+		log.Error(err)
+		return false, nil
+	}
+
+	success, resp, err := s.helixClient.Client.ValidateToken(token.AccessToken)
+	if !success || err != nil {
+		if err != nil {
+			log.Error(err)
+		}
+
+		// Token might be expired, let's try refreshing
+		if resp.Error == "Unauthorized" {
+			success, refreshResp := s.refreshToken(scToken, token)
+			if !success {
+				return false, nil
+			}
+
+			success, resp, err = s.helixClient.Client.ValidateToken(refreshResp.AccessToken)
+			if !success || err != nil {
+				if err != nil {
+					log.Error(err)
+				}
+
+				return success, resp
+			}
+		}
+
+		return false, nil
+	}
+
+	return success, resp
+}
+
+func (s *Server) refreshToken(scToken string, token userAcessTokenData) (bool, *userAcessTokenData) {
+	resp, err := s.helixClient.Client.RefreshUserAccessToken(token.RefreshToken)
+	if err != nil {
+		log.Error(err)
+		return false, nil
+	}
+
+	newToken := userAcessTokenData{resp.Data.AccessToken, resp.Data.RefreshToken, strings.Join(resp.Data.Scopes, " ")}
+	marshalled, err := json.Marshal(newToken)
+	if err != nil {
+		log.Error(err)
+		return false, nil
+	}
+
+	err = s.store.Client.HSet("accessTokens", scToken, marshalled).Err()
+	if err != nil {
+		log.Error(err)
+		return false, nil
+	}
+	log.Infof("Refreshed token for %s", scToken)
+
+	return true, &newToken
 }
